services/conversation_parser/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
black==25.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
distro==1.9.0
fastapi==0.121.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
isort==7.0.0
jiter==0.12.0
mypy_extensions==1.1.0
openai==2.8.1
packaging==25.0
pathspec==0.12.1
platformdirs==4.5.0
pydantic==2.12.4
pydantic_core==2.41.5
python-dotenv==1.2.1
pytokens==0.3.0
requests==2.32.5
ruff==0.14.6
sniffio==1.3.1
starlette==0.50.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
